@@ -5,9 +5,9 @@
 	<artifactId>flowing-retail-rest-stripe-fake</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<properties>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
-		<spring.boot.version>1.5.2.RELEASE</spring.boot.version>
+		<maven.compiler.source>17</maven.compiler.source>
+		<maven.compiler.target>17</maven.compiler.target>
+		<spring.boot.version>3.1.5</spring.boot.version>
 	</properties>
 	<dependencyManagement>
 		<dependencies>
