@@ -1,24 +1,39 @@
 package io.flowing.retail.payment;
 
+import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
+import io.github.resilience4j.core.registry.EntryAddedEvent;
+import io.github.resilience4j.core.registry.EntryRemovedEvent;
+import io.github.resilience4j.core.registry.EntryReplacedEvent;
+import io.github.resilience4j.core.registry.RegistryEventConsumer;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
-import io.zeebe.client.ZeebeClient;
-
 @SpringBootApplication
 public class PaymentRestApplication {
-
-  @Bean
-  public ZeebeClient zeebe() {
-    ZeebeClient zeebeClient = ZeebeClient.newClientBuilder() //
-    		.usePlaintext() // don't use SSL in default local broker
-    		.build();
-    return zeebeClient;
-  }
 
   public static void main(String[] args) throws Exception {
     SpringApplication.run(PaymentRestApplication.class, args);
   }
+  @Bean
+  public RegistryEventConsumer<CircuitBreaker> myRegistryEventConsumer() {
 
+    return new RegistryEventConsumer<CircuitBreaker>() {
+      @Override
+      public void onEntryAddedEvent(EntryAddedEvent<CircuitBreaker> entryAddedEvent) {
+        System.out.println("#################" + entryAddedEvent);
+        //entryAddedEvent.getAddedEntry().getEventPublisher().onEvent(event -> LOG.info(event.toString()));
+      }
+
+      @Override
+      public void onEntryRemovedEvent(EntryRemovedEvent<CircuitBreaker> entryRemoveEvent) {
+
+      }
+
+      @Override
+      public void onEntryReplacedEvent(EntryReplacedEvent<CircuitBreaker> entryReplacedEvent) {
+
+      }
+    };
+  }
 }
