@@ -4,18 +4,14 @@ import static org.springframework.web.bind.annotation.RequestMethod.PUT;
 
 import java.util.UUID;
 
-import javax.servlet.http.HttpServletResponse;
-
+import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
-import com.netflix.hystrix.HystrixCommand;
-import com.netflix.hystrix.HystrixCommandGroupKey;
-
 /**
- * Step2: Use Hystrix circuit breaker to make secured REST call
+ * Step2: Use circuit breaker to avoid cascading failures on slow REST calls
  */
 @RestController
 public class PaymentRestHacksControllerV2 {
@@ -25,7 +21,8 @@ public class PaymentRestHacksControllerV2 {
   private String stripeChargeUrl = "http://localhost:8099/charge";
   
   @RequestMapping(path = "/api/payment/v2", method = PUT)
-  public String retrievePayment(String retrievePaymentPayload, HttpServletResponse response) throws Exception {
+  @CircuitBreaker(name = "creditcard")
+  public String retrievePayment(String retrievePaymentPayload) throws Exception {
     String traceId = UUID.randomUUID().toString();
     String customerId = "0815"; // get somehow from retrievePaymentPayload
     long amount = 15; // get somehow from retrievePaymentPayload
@@ -38,19 +35,18 @@ public class PaymentRestHacksControllerV2 {
     CreateChargeRequest request = new CreateChargeRequest();
     request.amount = remainingAmount;
 
-    CreateChargeResponse response = new HystrixCommand<CreateChargeResponse>(HystrixCommandGroupKey.Factory.asKey("stripe")) {
-      protected CreateChargeResponse run() throws Exception {
-        return rest.postForObject( //
+    CreateChargeResponse response = rest.postForObject( //
             stripeChargeUrl, //
             request, //
             CreateChargeResponse.class);
-      }
-    }.execute();
-    
-    
+
     return response.transactionId;
   }
-  
+
+
+  private String fallback(String customerId, long remainingAmount, RuntimeException e) {
+    throw e;
+  }
   public static class CreateChargeRequest {
     public long amount;
   }
